Deleted .. 2 files, fixes on CupidMatch, Only delete vote if no remaining Votes left, refactored Selection #41

+++ src/components/ui/game/Selection.js
@@ -7,7 +7,7 @@ import storageManager from "../../../helpers/StorageManager";
 const Selection = ({currentPoll}) => {
 
   let voteParticipantIds = currentPoll.participants.map(p => p.player.id);
-
+  console.log("CurrentPoll: ", currentPoll);
   const singleVoters = ["Witch-Kill", "Hunter", "Seer", "Mayor"];
   const selectionSize = singleVoters.includes(currentPoll.role) ? "selection-big" : "selection-small";
   
@@ -17,7 +17,9 @@ const Selection = ({currentPoll}) => {
       if(currentPoll.ownVote == null) {
         await api.put("/games/" + storageManager.getLobbyId() + "/votes/" + optionId);
       } else {
-        await api.delete("/games/" + storageManager.getLobbyId() + "/votes/" + currentPoll.ownVote.id);
+        if (currentPoll.getOwnRemainingVotes() == 0) {
+          await api.delete("/games/" + storageManager.getLobbyId() + "/votes/" + currentPoll.ownVote.id);
+        }
         await api.put("/games/" + storageManager.getLobbyId() + "/votes/" + optionId);
       }
     } catch (error) {
@@ -26,49 +28,41 @@ const Selection = ({currentPoll}) => {
     }
   };
 
+  const selectionProfiles = (playerCollection)  => {
+    console.log("Player Collection: ", playerCollection);
+    return (
+      <div className="game-player-selection-wrapper">
+        {playerCollection.map(option => (
+          <Profile user={new Player(option.player)} mode={selectionSize} onClickEvent={castVote} key={option.player.id} />
+          // <Profile user={new Player(option.player)} mode={selectionSize} onClickEvent={castVote} key={option.id} />
+        ))}
+      </div>
+    );
+  };
+
   let content;
  
   switch (currentPoll.role) {
     case 'Werewolf':
       content = (
         <>
-        <div className="game-player-selection-wrapper">
-          {currentPoll.pollOptions.map(option => (
-            (!voteParticipantIds.includes(option.player.id)) && (
-              <Profile user={new Player(option.player)} mode="selection-small" onClickEvent={castVote} key={option.player.id} />
-            )
-          ))}
-        </div>
+        {selectionProfiles(currentPoll.pollOptions.filter(option => !voteParticipantIds.includes(option.player.id)))}
         <h2>Your fellow werewolves:</h2>
-        <div className="game-player-selection-wrapper">
-          {currentPoll.pollOptions.map(option => (
-            (voteParticipantIds.includes(option.player.id)) && (
-              <Profile user={new Player(option.player)} mode="selection-small" onClickEvent={castVote} key={option.player.id} />
-            )
-          ))}
-        </div>
+        {selectionProfiles(currentPoll.pollOptions.filter(option => voteParticipantIds.includes(option.player.id)))}
         </>
       );
       break;
     default:
-        content = (
-          <>
-            <div className="game-player-selection-wrapper">
-              {currentPoll.pollOptions.map(option => {
-                return (
-                  <Profile user={new Player(option.player)} mode={selectionSize} onClickEvent={castVote} key={option.player.id} />
-                );
-              })}
-            </div>
-          </>
-        );
+      content = (
+        selectionProfiles(currentPoll.pollOptions)
+      );
       break;
   }
 
   return (
     <div className={`game-player-selection ${currentPoll.participants.length > 0 ? "game-player-selection-active": ""}`}>
       {content}
-      </div>
+    </div>
   );
   
 }
